Completed html structure & SCSS fixes
This basic layout of the tic-tac-toe game is fully structured is pretty
enough to move forward;
Added:
- an additional fieldset ( sign-out ) to the index.html page
- an additional SCSS file (typography.scss) to the styles dir
  > this contains the only CDN font needed for this project

The webpage contained an awful amount of negative space that made the
game unappealing. To fix this:
header position is now static
The game board and user forms were centered and repositioned
The game board squares are now 150px^2

*For Next Commit / Current Bugs*
- The form submit buttons refresh the page => must prevent default!
- BEFORE adding main event listeners to the game board,
an event listener must be attached to the 'PLAY GAME' button.
This should be able to toggle to display attribute of my game board.