Started Event Handling & MORE SCSS
Begun creating the event handling logic for the website
Added:
- api.js
- events.js
- ui.js

Inside events.js, the first onSignup event was created and I copied the
api.js function from the one made previously in class.

Insde app.js, the click event was created and attached on to the
'draw-board' button. The event listener will toggle the display of the
game board.

*For Next Commit / Current Bugs*
- the footer tag and all of its contents have a hard time staying at the
bottom of the page
  > this is presumably a style issue not a priority at the moment

- the api response div will likely need to be repositioned to fit
  a reasonably sized string of text as well as the game board.

>> Moving Forward with game logic
>> User must be able to click on a cell and place their letter {x / o}

- After may start a new branch and experiment with two different routes
- the most important aspect is making sure the game engine handles the
  game data in a way the API expects.
- I plan on creating a game object in the store.js that will hold the
current game data and update the gameAPI from there.. Checks out right?

lets hope so