Added Game Logic - Storing Player Moves
Added:
- storePlayerMove() Game method
- checkForMatch() event
- checkWin() event

Completed the logic for storing the moves of each player and placed it
inside the Game class to use as method.

Began creating the logic to resolve the game once it has begun.
For now the checkForMatch function contains $variables of the rows and
columns of the board. A for each loop will be used to use checkWin
for winning combinations.

-> At the moment I do not have much confidence in this approach above
  and I will play around with my logic a bit.
-> Overall the checkForMatch function will need to be rewritten probably

*For Next Commit / Current Bugs*
- Game does not end
- Game does not have a clear winner
- The API still does not know what the players are doing