Completed Game logic ...... for now
Added:
- jquery click event in app.js > onPlayAgain
- onPlayAgain() function in event.js
- A 'Play Again' button inside the Main html tag
-> Fixed a bug where the currentGame ID was not updating correctly
-> removed a lot of console logs I did not need
-> Moved around lines of code

After a few failed implementations of creating a reset button and/or
trying to add a conditional inside the drawBoard function, I created a
new click event.

Most of the code that I needed was already apart of the program. The end
game function now only adds the endGame class and toggles the new
'play again' button.

onPlayAgain() will now reset the board, remove the endGame class, hide
the 'play again' button, and invoke the newGame() method that is in the
tic-tac-toe Game class.

*For Next Commit/ No Bugs *
- API communication
- Game class should be able to update the game moves
- log out button can only be visible when the user is logged in
- Ui needs to update Dom when the user is signing in